add detailed test to make sure the config, env.make uses the same kwargs name
